+++ aie/data/sticker.cpp
