+++ hw_link/config.cfg
@@ -1,0 +1,145 @@
+[connectivity]
+# Kernels
+# nk=mm2s:32:mm2s1.mm2s2.mm2s3.mm2s4.mm2s5.mm2s6.mm2s7.mm2s8.mm2s9.mm2s10.mm2s11.mm2s12.mm2s13.mm2s14.mm2s15.mm2s16.mm2s17.mm2s18.mm2s19.mm2s20.mm2s21.mm2s22.mm2s23.mm2s24.mm2s25.mm2s26.mm2s27.mm2s28.mm2s29.mm2s30.mm2s31.mm2s32
+# nk=s2mm:32:s2mm1.s2mm2.s2mm3.s2mm4.s2mm5.s2mm6.s2mm7.s2mm8.s2mm9.s2mm10.s2mm11.s2mm12.s2mm13.s2mm14.s2mm15.s2mm16.s2mm17.s2mm18.s2mm19.s2mm20.s2mm21.s2mm22.s2mm23.s2mm24.s2mm25.s2mm26.s2mm27.s2mm28.s2mm29.s2mm30.s2mm31.s2mm32
+nk=mm2s:15:mm2s1.mm2s2.mm2s3.mm2s4.mm2s5.mm2s6.mm2s7.mm2s8.mm2s9.mm2s10.mm2s11.mm2s12.mm2s13.mm2s14.mm2s15
+nk=s2mm:15:s2mm1.s2mm2.s2mm3.s2mm4.s2mm5.s2mm6.s2mm7.s2mm8.s2mm9.s2mm10.s2mm11.s2mm12.s2mm13.s2mm14.s2mm15
+
+
+# DDR to AIE
+stream_connect=mm2s1.s:ai_engine_0.DataIn1
+stream_connect=mm2s2.s:ai_engine_0.DataIn2
+stream_connect=mm2s3.s:ai_engine_0.DataIn3
+stream_connect=mm2s4.s:ai_engine_0.DataIn4
+stream_connect=mm2s5.s:ai_engine_0.DataIn5
+stream_connect=mm2s6.s:ai_engine_0.DataIn6
+stream_connect=mm2s7.s:ai_engine_0.DataIn7
+stream_connect=mm2s8.s:ai_engine_0.DataIn8
+stream_connect=mm2s9.s:ai_engine_0.DataIn9
+stream_connect=mm2s10.s:ai_engine_0.DataIn10
+stream_connect=mm2s11.s:ai_engine_0.DataIn11
+stream_connect=mm2s12.s:ai_engine_0.DataIn12
+stream_connect=mm2s13.s:ai_engine_0.DataIn13
+stream_connect=mm2s14.s:ai_engine_0.DataIn14
+stream_connect=mm2s15.s:ai_engine_0.DataIn15
+# stream_connect=mm2s16.s:ai_engine_0.DataIn16
+# stream_connect=mm2s17.s:ai_engine_0.DataIn17
+# stream_connect=mm2s18.s:ai_engine_0.DataIn18
+# stream_connect=mm2s19.s:ai_engine_0.DataIn19
+# stream_connect=mm2s20.s:ai_engine_0.DataIn20
+# stream_connect=mm2s21.s:ai_engine_0.DataIn21
+# stream_connect=mm2s22.s:ai_engine_0.DataIn22
+# stream_connect=mm2s23.s:ai_engine_0.DataIn23
+# stream_connect=mm2s24.s:ai_engine_0.DataIn24
+# stream_connect=mm2s25.s:ai_engine_0.DataIn25
+# stream_connect=mm2s26.s:ai_engine_0.DataIn26
+# stream_connect=mm2s27.s:ai_engine_0.DataIn27
+# stream_connect=mm2s28.s:ai_engine_0.DataIn28
+# stream_connect=mm2s29.s:ai_engine_0.DataIn29
+# stream_connect=mm2s30.s:ai_engine_0.DataIn30
+# stream_connect=mm2s31.s:ai_engine_0.DataIn31
+# stream_connect=mm2s32.s:ai_engine_0.DataIn32
+
+# AIE to DDR
+stream_connect=ai_engine_0.DataOut1:s2mm1.s
+stream_connect=ai_engine_0.DataOut2:s2mm2.s
+stream_connect=ai_engine_0.DataOut3:s2mm3.s
+stream_connect=ai_engine_0.DataOut4:s2mm4.s
+stream_connect=ai_engine_0.DataOut5:s2mm5.s
+stream_connect=ai_engine_0.DataOut6:s2mm6.s
+stream_connect=ai_engine_0.DataOut7:s2mm7.s
+stream_connect=ai_engine_0.DataOut8:s2mm8.s
+stream_connect=ai_engine_0.DataOut9:s2mm9.s
+stream_connect=ai_engine_0.DataOut10:s2mm10.s
+stream_connect=ai_engine_0.DataOut11:s2mm11.s
+stream_connect=ai_engine_0.DataOut12:s2mm12.s
+stream_connect=ai_engine_0.DataOut13:s2mm13.s
+stream_connect=ai_engine_0.DataOut14:s2mm14.s
+stream_connect=ai_engine_0.DataOut15:s2mm15.s
+# stream_connect=ai_engine_0.DataOut16:s2mm16.s
+# stream_connect=ai_engine_0.DataOut17:s2mm17.s
+# stream_connect=ai_engine_0.DataOut18:s2mm18.s
+# stream_connect=ai_engine_0.DataOut19:s2mm19.s
+# stream_connect=ai_engine_0.DataOut20:s2mm20.s
+# stream_connect=ai_engine_0.DataOut21:s2mm21.s
+# stream_connect=ai_engine_0.DataOut22:s2mm22.s
+# stream_connect=ai_engine_0.DataOut23:s2mm23.s
+# stream_connect=ai_engine_0.DataOut24:s2mm24.s
+# stream_connect=ai_engine_0.DataOut25:s2mm25.s
+# stream_connect=ai_engine_0.DataOut26:s2mm26.s
+# stream_connect=ai_engine_0.DataOut27:s2mm27.s
+# stream_connect=ai_engine_0.DataOut28:s2mm28.s
+# stream_connect=ai_engine_0.DataOut29:s2mm29.s
+# stream_connect=ai_engine_0.DataOut30:s2mm30.s
+# stream_connect=ai_engine_0.DataOut31:s2mm31.s
+# stream_connect=ai_engine_0.DataOut32:s2mm32.s
+
+# sp=mm2s1.s:DDR[0]
+# sp=mm2s2.s:DDR[0]
+# sp=mm2s3.s:DDR[0]
+# sp=mm2s4.s:DDR[0]
+# sp=mm2s5.s:DDR[0]
+# sp=mm2s6.s:DDR[0]
+# sp=mm2s7.s:DDR[0]
+# sp=mm2s8.s:DDR[0]
+# sp=mm2s9.s:DDR[0]
+# sp=mm2s10.s:DDR[0]
+# sp=mm2s11.s:DDR[0]
+# sp=mm2s12.s:DDR[0]
+# sp=mm2s13.s:DDR[0]
+# sp=mm2s14.s:DDR[0]
+# sp=mm2s15.s:DDR[0]
+# sp=mm2s16.s:DDR[1]
+# sp=mm2s17.s:DDR[1]
+# sp=mm2s18.s:DDR[1]
+# sp=mm2s19.s:DDR[1]
+# sp=mm2s20.s:DDR[1]
+# sp=mm2s21.s:DDR[1]
+# sp=mm2s22.s:DDR[1]
+# sp=mm2s23.s:DDR[1]
+# sp=mm2s24.s:DDR[1]
+# sp=mm2s25.s:DDR[1]
+# sp=mm2s26.s:DDR[1]
+# sp=mm2s27.s:DDR[1]
+# sp=mm2s28.s:DDR[1]
+# sp=mm2s29.s:DDR[1]
+# sp=mm2s30.s:DDR[1]
+
+# sp=s2mm1.s:DDR[2]
+# sp=s2mm2.s:DDR[2]
+# sp=s2mm3.s:DDR[2]
+# sp=s2mm4.s:DDR[2]
+# sp=s2mm5.s:DDR[2]
+# sp=s2mm6.s:DDR[2]
+# sp=s2mm7.s:DDR[2]
+# sp=s2mm8.s:DDR[2]
+# sp=s2mm9.s:DDR[2]
+# sp=s2mm10.s:DDR[2]
+# sp=s2mm11.s:DDR[2]
+# sp=s2mm12.s:DDR[2]
+# sp=s2mm13.s:DDR[2]
+# sp=s2mm14.s:DDR[2]
+# sp=s2mm15.s:DDR[2]
+# sp=s2mm16.s:DDR[3]
+# sp=s2mm17.s:DDR[3]
+# sp=s2mm18.s:DDR[3]
+# sp=s2mm19.s:DDR[3]
+# sp=s2mm20.s:DDR[3]
+# sp=s2mm21.s:DDR[3]
+# sp=s2mm22.s:DDR[3]
+# sp=s2mm23.s:DDR[3]
+# sp=s2mm24.s:DDR[3]
+# sp=s2mm25.s:DDR[3]
+# sp=s2mm26.s:DDR[3]
+# sp=s2mm27.s:DDR[3]
+# sp=s2mm28.s:DDR[3]
+# sp=s2mm29.s:DDR[3]
+# sp=s2mm30.s:DDR[3]
+
+
+
+[advanced]
+param=compiler.errorOnHoldViolation=false
+
+[vivado]
+prop=run.impl_1.strategy=Performance_NetDelay_low
