+++ pl/src/config.h
@@ -1,0 +1,10 @@
+// Copyright (C) 2023 Advanced Micro Devices, Inc
+//
+// SPDX-License-Identifier: MIT
+
+#include <ap_axi_sdata.h>
+#include <ap_int.h>
+#include <hls_stream.h>
+
+#define DWIDTH 32
+typedef qdma_axis<DWIDTH, 0, 0, 0> data;
